Creation of fileHandling.py
